Ajuste para uso de BIGINT no SQL - Para que funcione corretamente com o Long no Java

+++ src/main/resources/db/migration/V2__Create-Table-enderecos.sql
@@ -1,11 +1,11 @@
 CREATE TABLE enderecos (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     ativo BOOLEAN DEFAULT TRUE,
     rua VARCHAR(255),
     numero VARCHAR(10),
     bairro VARCHAR(255),
     cidade VARCHAR(255),
     estado VARCHAR(2),
-    usuario_id INT UNIQUE, -- Relacionamento um-para-um com usuario
+    usuario_id BIGINT UNIQUE, -- Relacionamento um-para-um com usuario
     FOREIGN KEY (usuario_id) REFERENCES usuarios (id)
 );
