+++ src/main/resources/db/migration/V5__Create-Table-usuario_roles.sql
@@ -1,5 +1,5 @@
 CREATE TABLE usuario_roles (
-    usuario_id INT NOT NULL,
+    usuario_id BIGINT NOT NULL,
     role_id INT NOT NULL,
     PRIMARY KEY (usuario_id, role_id),
     FOREIGN KEY (usuario_id) REFERENCES usuarios (id) ON DELETE CASCADE,
