+++ src/main/resources/db/migration/V9__Create-Table-doacoes.sql
@@ -1,9 +1,9 @@
 CREATE TABLE doacoes (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     valor NUMERIC(10, 2) NOT NULL,
     data TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    id_doador INT NOT NULL, -- Quem está doando
-    id_donatario INT NOT NULL, -- Quem está recebendo a doação
+    id_doador BIGINT NOT NULL, -- Quem está doando
+    id_donatario BIGINT NOT NULL, -- Quem está recebendo a doação
     FOREIGN KEY (id_doador) REFERENCES usuarios (id),
     FOREIGN KEY (id_donatario) REFERENCES usuarios (id)
 );
